Extract javascript for testability; set up mocha.
