+++ package.json
@@ -1,0 +1,25 @@
+{
+  "name": "ziggy",
+  "version": "0.2.3",
+  "description": "Generates a Blade directive exporting all of your named Laravel routes. Also provides a nice route() helper function in JavaScript.",
+  "main": "src/js/route.js",
+  "directories": {
+    "test": "tests/js"
+  },
+  "scripts": {
+    "test": "mocha 'tests/js'"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/tightenco/ziggy.git"
+  },
+  "author": "Daniel Coulbourne",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/tightenco/ziggy/issues"
+  },
+  "homepage": "https://github.com/tightenco/ziggy#readme",
+  "devDependencies": {
+    "mocha": "^3.5.0"
+  }
+}
