+++ src/BladeRouteGenerator.php
@@ -18,27 +18,13 @@ class BladeRouteGenerator
     {
         $json = (string) $this->nameKeyedRoutes();
         $appUrl = rtrim(config('app.url'), '/') . '/';
+        $routeFunction = file_get_contents(__DIR__ . '/js/route.js');
 
         return <<<EOT
 <script type="text/javascript">
     var namedRoutes = JSON.parse('$json'),
         baseUrl = '$appUrl';
-
-    function route (name, params = {}, absolute = true) {
-        var domain = namedRoutes[name].domain || baseUrl,
-            url = (absolute ? domain : '') + namedRoutes[name].uri
-
-        return url.replace(
-            /\{([^}]+)\}/gi,
-            function (tag) {
-                var key = tag.replace(/\{|\}/gi, '');
-                if (params[key] === undefined) {
-                    throw 'Ziggy Error: "' + key + '" key is required for route "' + name + '"';
-                }
-                return params[key];
-            }
-        );
-    }
+    $routeFunction
 </script>
 EOT;
     }
